Allow reconnection within 20 seconds
The last room is now stored in sessionStorage. If you refresh the
page, it will first try to rejoin the room using the last
sessionId. In this way you will keep your hand if you stay on the same
page and reconnect within 20 seconds.

The downside is that people who leave will stay in the room for 20 seconds.